Update index.html
a whole lot of SEO optimization.
Unfortunately, you use H1 and whatnot for a lot of these so its kinda hard for me to go in and edit those easily because H1's and whatnot are typically massive SEO indicators so i did my best to edit non-essential stuff. I slightly sized down one part of the FAQ and i also edited alot of the text on the page for SEO keyword optimization.